cleanup, fix tests
