+++ network-api/networkapi/buyersguide/views.py
@@ -15,7 +15,7 @@ from rest_framework.response import Response
 
 from networkapi.buyersguide.models import Product, BuyersGuideProductCategory, BooleanVote, RangeVote
 from networkapi.buyersguide.throttle import UserVoteRateThrottle, TestUserVoteRateThrottle
-from networkapi.buyersguide.utils import redirect_to_main_cms_site_decorator
+from networkapi.utility.redirects import redirect_to_default_cms_site_decorator
 
 vote_throttle_class = UserVoteRateThrottle if not settings.TESTING else TestUserVoteRateThrottle
 
@@ -68,7 +68,7 @@ def filter_draft_products(request, products):
     return filter(lambda p: p['draft'] is False, products)
 
 
-@redirect_to_main_cms_site_decorator
+@redirect_to_default_cms_site_decorator
 @enforce_en_locale
 def buyersguide_home(request):
     products = cache.get('sorted_product_dicts')
@@ -87,7 +87,7 @@ def buyersguide_home(request):
     })
 
 
-@redirect_to_main_cms_site_decorator
+@redirect_to_default_cms_site_decorator
 @enforce_en_locale
 def category_view(request, slug):
     key = f'products_category__{slug}'
@@ -113,7 +113,7 @@ def category_view(request, slug):
     })
 
 
-@redirect_to_main_cms_site_decorator
+@redirect_to_default_cms_site_decorator
 @enforce_en_locale
 def product_view(request, slug):
     product = get_object_or_404(Product, slug=slug)
@@ -151,7 +151,7 @@ def product_view(request, slug):
 
 
 def bg_about_page(template_name):
-    @redirect_to_main_cms_site_decorator
+    @redirect_to_default_cms_site_decorator
     @enforce_en_locale
     def render_view(request):
         key = 'categories'
